requirements.txt
fastapi
uvicorn
pyngrok
openai-whisper
python-multipart
